+++ modules/services/improveOSM.js
@@ -57,11 +57,11 @@ function updateRtree(item, replace) {
 }
 
 function linkErrorObject(d) {
-  return `<a class="error_object_link">${d}</a>`;
+  return { html: `<a class="error_object_link">${d}</a>` };
 }
 
 function linkEntity(d) {
-  return `<a class="error_entity_link">${d}</a>`;
+  return { html: `<a class="error_entity_link">${d}</a>` };
 }
 
 function pointAverage(points) {