+++ modules/ui/improveOSM_details.js
@@ -15,7 +15,7 @@ export function uiImproveOsmDetails(context) {
     if (d.desc) return d.desc;
     const issueKey = d.issueKey;
     d.replacements = d.replacements || {};
-    d.replacements.default = t.html('inspector.unknown');  // special key `default` works as a fallback string
+    d.replacements.default = { html: t.html('inspector.unknown') };  // special key `default` works as a fallback string
     return t.html(`QA.improveOSM.error_types.${issueKey}.description`, d.replacements);
   }
 
@@ -47,7 +47,7 @@ export function uiImproveOsmDetails(context) {
     descriptionEnter
       .append('div')
         .attr('class', 'qa-details-description-text')
-        .text(issueDetail);
+        .html(issueDetail);
 
     // If there are entity links in the error message..
     let relatedEntities = [];